requirements.txt
asn1crypto==0.24.0
cffi==1.12.1
Click==7.0
cryptography==2.5
enum34==1.1.6
Flask==1.0.2
Flask-JWT-Extended==3.17.0
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector-python==8.0.15
protobuf==3.6.1
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
waitress==1.2.1
Werkzeug==0.14.1
